source/requirements.txt
deep-translator
progressbar2